Int-Parameter: Float-Werte vor dem Mappen nicht mehr abschneiden
Der 'f'-Zweig in RemoteControlledIntParameter.digestMessage las den
normierten Wert mit intValue() statt floatValue() - also VOR der
Spreizung von 0..1 auf min..max. Damit wurde 0.75 zu 0 und landete auf
minValue; nur exakt 1.0 erreichte das Maximum.

Sichtbar wird das bei /net/impulse/oscMaxCount (int, 0..256): ein
normierter Fader stellt de facto 0 ein, die Klangschicht verstummt,
waehrend /net/hitNode weiter feuert und die Oberflaeche reagiert - ein
Ausfall, der wie funktionierende Software aussieht. Ebenso betroffen
sind /net/randomSpawn/count und /net/randomSpawn/enabled.

Der 'i'-Zweig bleibt unangetastet: Sender, die bereits Ints schicken,
kommen an dieser Stelle gar nicht vorbei. RemoteControlledFloatParameter
macht es an der gleichen Stelle schon immer richtig - das hier war eine
Kopier-Ungenauigkeit.

+++ AbstractParameter.java
@@ -180,7 +180,13 @@ class RemoteControlledIntParameter extends IntParameter implements OscMessageSin
 				setValue((int) theValue);
 			}
 			if (newMessage.getTypetagAsBytes()[0] == 'f') {
-				float theValue = newMessage.get(0).intValue();
+				// floatValue(), nicht intValue(): der Wert kommt normiert als 0..1 herein
+				// und wird erst hier auf min..max gespreizt. Mit intValue() wurde vor dem
+				// Mappen abgeschnitten - 0.75 wurde zu 0 und landete auf minValue, nur
+				// exakt 1.0 erreichte das Maximum. Bei /net/impulse/oscMaxCount hiess das
+				// 0 statt 192: die Klangschicht verstummt, waehrend die Oberflaeche
+				// reagiert und /net/hitNode weiter feuert.
+				float theValue = newMessage.get(0).floatValue();
 				theValue = PApplet.constrain(PApplet.map(theValue, 0f, 1f, minValue, maxValue), minValue, maxValue);
 				setValue((int) theValue);
 			}
